+++ controller/listCtrl.js
@@ -9,6 +9,7 @@ async function createToDoList(req, res) {
   if (access === null) {
     return res.json({ loginStatus: false });
   }
+
   try {
     const createResult = await listModel.createList(access.userId, listTitle);
     if (!createResult) {
@@ -27,6 +28,7 @@ async function updatedToDoList(req, res) {
   const { listId, listTitle } = req.body;
   const token = req.header.Authorization;
   const access = tools.verifyToken(token);
+  const checkPass = tools.checkUserId(listId);
   if (access === null) {
     return res.json({ loginStatus: false });
   }
@@ -35,6 +37,10 @@ async function updatedToDoList(req, res) {
       .status(200)
       .json({ updatedList: false, message: '輸入非正整數型別' });
   }
+  if (checkPass != access.userId || !checkPass) {
+    return res.status(200).json({ updatedList: false, message: '無此清單' });
+  }
+
   try {
     const updateResult = await listModel.updatedList(listId, listTitle);
     if (!updateResult) {
@@ -57,6 +63,7 @@ async function deleteToDoList(req, res) {
   const listId = req.body.listId;
   const token = req.header.Authorization;
   const access = tools.verifyToken(token);
+  const checkPass = tools.checkUserId(listId);
   const areAllNumbers = listId.every((item) => typeof item === 'number');
   if (access === null) {
     return res.json({ loginStatus: false });
@@ -65,6 +72,9 @@ async function deleteToDoList(req, res) {
     return res
       .status(200)
       .json({ removeList: false, message: '輸入非正整數型別' });
+  }
+  if (checkPass != access.userId || !checkPass) {
+    return res.status(200).json({ removeList: false, message: '無此清單' });
   }
   try {
     const deleteResult = await listModel.deleteList(listId, access.userId);
